test: verify auditlog work and organise classes
