+++ src/test/kotlin/mockdata/CreateUserHelper.kt
@@ -1,0 +1,13 @@
+package mockdata
+
+import org.example.logic.models.User
+import org.example.logic.models.UserRole
+
+fun createUser(
+    role: UserRole
+) = User(
+    id = "1",
+    username = "",
+    password = "",
+    role = role
+)
